Move VSBranchInfo to the dotnet tool
