+++ RoslynTools.sln
@@ -68,8 +68,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PRFinder", "src\RoslynInser
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CreateTagsForVSRelease", "src\CreateTagsForVSRelease\CreateTagsForVSRelease.csproj", "{08AAEF36-C81B-4BAE-8172-EA0D3139B0BE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VSBranchInfo", "src\VSBranchInfo\VSBranchInfo.csproj", "{D10CCD63-A7F6-4310-81AA-0D00526320D1}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NugetDependencyFinder", "src\NugetDependencyFinder\NugetDependencyFinder.csproj", "{7A6963D6-1F40-4CC6-9941-7F912252F877}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "dotnet-roslyn", "dotnet-roslyn", "{C68033E0-FB8F-4327-BFB5-B340A5A7778F}"
@@ -362,18 +360,6 @@ Global
 		{08AAEF36-C81B-4BAE-8172-EA0D3139B0BE}.Release|x64.Build.0 = Release|Any CPU
 		{08AAEF36-C81B-4BAE-8172-EA0D3139B0BE}.Release|x86.ActiveCfg = Release|Any CPU
 		{08AAEF36-C81B-4BAE-8172-EA0D3139B0BE}.Release|x86.Build.0 = Release|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Debug|x64.Build.0 = Debug|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Debug|x86.Build.0 = Debug|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Release|x64.ActiveCfg = Release|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Release|x64.Build.0 = Release|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Release|x86.ActiveCfg = Release|Any CPU
-		{D10CCD63-A7F6-4310-81AA-0D00526320D1}.Release|x86.Build.0 = Release|Any CPU
 		{7A6963D6-1F40-4CC6-9941-7F912252F877}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{7A6963D6-1F40-4CC6-9941-7F912252F877}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{7A6963D6-1F40-4CC6-9941-7F912252F877}.Debug|x64.ActiveCfg = Debug|Any CPU
