--- src/VSBranchInfo/VSBranchInfo.csproj
@@ -1,17 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.5.0-beta.1" />
-    <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.2.0-beta.5" />
-    <PackageReference Include="Microsoft.TeamFoundationServer.Client" Version="16.185.0-preview" />
-    <PackageReference Include="Microsoft.VisualStudio.Services.Client" Version="16.185.0-preview" />
-    <PackageReference Include="System.Net.Http" Version="$(SystemNetHttpVersion)" />
-  </ItemGroup>
-
-</Project>
