+++ src/dotnet-roslyn/Microsoft.Roslyn.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -22,6 +22,12 @@
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta2.21615.1" />
     <PackageReference Include="System.CommandLine.Rendering" Version="0.4.0-alpha.21615.1" />
     <PackageReference Include="Microsoft.Extensions.Logging" Version="6.0.0" />
+    
+    <PackageReference Include="Azure.Identity" Version="1.6.0-beta.1" />
+    <PackageReference Include="Azure.Security.KeyVault.Secrets" Version="4.3.0-beta.4" />
+    <PackageReference Include="Microsoft.TeamFoundationServer.Client" Version="16.201.0-preview" />
+    <PackageReference Include="Microsoft.VisualStudio.Services.Client" Version="16.201.0-preview" />
+    <PackageReference Include="System.Net.Http" Version="$(SystemNetHttpVersion)" />
   </ItemGroup>
 
 </Project>
